requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
jinja2==3.1.2
pydantic-settings==2.1.0
apscheduler==3.10.4
psutil==5.9.6
aiofiles==23.2.1